El componente del mapa lo divide en mas componentes, separe la logica del mapa y tambien separe el proxy para el mapa y asi mantener escondido la llave
