+++ next.config.ts
@@ -5,7 +5,7 @@ const nextConfig: NextConfig = {
     remotePatterns: [
       {
         protocol: 'https',
-        hostname: 'lh3.googleusercontent.com',
+        hostname: 'proveedores.onrentx.com',
         pathname: '/**',
       },
       {
